Backing up project
Converting the project to use threading, as opposed to asyncio. This change will better suite the needs of the project, and relieve some complexity in the functions.
As it stands, there are many synchronous functions that only exist to run async code. That's dirty. With threading, we only need to use a thread where we need concurrency, and allow the rest of the program to run as normal.